+++ DearPyGui/src/Core/PlotAppItems/mvAreaSeries.cpp
@@ -9,11 +9,11 @@ namespace Marvel {
 		std::vector<ImVec2> points;
 		for (unsigned i = 0; i < m_xs.size(); i++)
 		{
-			float x = m_xs[i] > limits.X.Max ? limits.X.Max : m_xs[i];
-			x = m_xs[i] < limits.X.Min ? limits.X.Min : x;
+			float x = m_xs[i] > limits.X.Max ? (float)limits.X.Max : (float)m_xs[i];
+			x = m_xs[i] < limits.X.Min ? (float)limits.X.Min : x;
 
-			float y = m_ys[i] > limits.Y.Max ? limits.Y.Max : m_ys[i];
-			y = m_ys[i] < limits.Y.Min ? limits.Y.Min : y;
+			float y = m_ys[i] > limits.Y.Max ? (float)limits.Y.Max : (float)m_ys[i];
+			y = m_ys[i] < limits.Y.Min ? (float)limits.Y.Min : y;
 			auto p = ImPlot::PlotToPixels({ x, y });
 			points.push_back(p);
 		}
@@ -50,8 +50,8 @@ namespace Marvel {
 					}
 					else
 					{
-						ind1 = i - 1;
-						ind2 = i;
+						ind1 = (int)i - 1;
+						ind2 = (int)i;
 					}
 					y1 = (int)points[ind1].y;
 					y2 = (int)points[ind2].y;
