+++ DearPyGui/src/Core/PlotAppItems/mvErrorSeries.h
@@ -19,9 +19,9 @@ namespace Marvel {
 		{
 			ImPlot::PushStyleColor(ImPlotCol_ErrorBar, m_color.toVec4());
 			if(m_horizontal)
-				ImPlot::PlotErrorBarsH(m_name.c_str(), m_xs.data(), m_ys.data(), m_extra1.data(), m_extra2.data(), m_xs.size());
+				ImPlot::PlotErrorBarsH(m_name.c_str(), m_xs.data(), m_ys.data(), m_extra1.data(), m_extra2.data(), (int)m_xs.size());
 			else
-				ImPlot::PlotErrorBars(m_name.c_str(), m_xs.data(), m_ys.data(), m_extra1.data(), m_extra2.data(), m_xs.size());
+				ImPlot::PlotErrorBars(m_name.c_str(), m_xs.data(), m_ys.data(), m_extra1.data(), m_extra2.data(), (int)m_xs.size());
 
 			ImPlot::PopStyleColor();
 		}
